refactor: move NavigateInput to widgets dir
