+++ src/components/Browse.tsx
@@ -14,7 +14,7 @@ import ChangePermissions from './ui/Dialogs/ChangePermissions';
 import ConvertFileDialog from './ui/Dialogs/ConvertFile';
 import RecentDataLinksCard from './ui/BrowsePage/Dashboard/RecentDataLinksCard';
 import RecentlyViewedCard from './ui/BrowsePage/Dashboard/RecentlyViewedCard';
-import NavigationInput from './NavigateInput';
+import NavigationInput from './ui/widgets/NavigateInput';
 import NavigationDialog from './ui/Dialogs/Navigate';
 
 type OutletContextType = {
